refactor(cli): extract theme/analyze/health into commands/diagnostics_commands.py
Moves the three diagnostic commands from cli.py into a dedicated module:
- `theme` for ritual aesthetic toggling
- `analyze` for high-fidelity codebase analysis
- `health` for the full ecosystem diagnostic HUD (including cleanup,
  watch, fix, and service-targeted modes)

No behavior changes. cli.py drops ~348 lines; diagnostics_commands.py
is self-contained with its own lazy imports for DocumentProcessor.
